msgb: Optionally declare some msgb struct fields as const
Writing directly to following struct fields may cause inconsistencies
that are hard to debug:

  data_len, len, head, tail, data

In general, the available macros and functions should be used to
modify them instead.

This patch declares these fields as const if
MSGB_DISABLE_DIRECT_WRITE is defined. Doing so may lead to warnings
and errors, therefore this macro is only defined for libosmocore yet,
where at least the errors are also fixed by this patch.

The main purpose is to maintain consistency, so only modifing the
fields themselves is restricted. It's still possible to modify the
data the pointers refer to.

Sponsored-by: On-Waves ehf