Integrate asynchronous DNS resolver (c-ares) with libosmocore
As osmocom-style applications are centered around a single select loop,
DNS queries need to be handled asynchronously to prevent blocking the
entire program.

The c-ares library is a popular option for async DNS resolving.
However, it requires integration with the select loop of the
application, which in our case lives inside libosmocore select.c.

This patch adds the required integration.  The backside is that it adds
a mandatory dependency to c-ares (libcares), as well as 2-3 additional
branches to every call to osmo_select_main().