logging: Remember the target we found
log_check_level and osmo_vlogp share the responsibility for the
output of a log message. Once check_level has identified the first
target that might have an output osmo_vlogp will continue from this
place.

In practice we have one (stderr/syslog) or two (stderr+VTY) log
outputs so avoiding to re-iterate is not that important but as
we have found the right place we can just use it.