libosmogb: Support multiple instances also from VTY
The NS code (part of libosmogb) already understands the concept of "NS
instances", and each of those instances can be bound to different IP
addresses with the gprs_ns_nsip_listen() call.

The problem is the VTY code for NS. It is written with a single
global ns-instance in mind. Having a second NS instance thus means that
currently only one of the two would be configurable (or available for
"show..." commands) from the VTY, which is of course of little
usefulness.

This patch addresses this by introducing the concept of "NS instance
numbers" and a global array of a maximum of 256 NS instances. The VTY
commands have been extended as needed.

Change-Id: I483e9984f046e0e698c65b59034b3a6141e7ca73