doxygen: gsm: do not exclude include/gsm/protocol/*
Clear out the EXCLUDE tag to not exclude anything.

Tried to add API doc to gsm/protocol/gsm_08_08.h and after a long time found
out why they are not appearing in the API doc.

The original reason to exclude these files was that we were not going to
copy-paste descriptions from the 3GPP specs. Nevertheless, for an API doc
reader it is helpful if all data structures are listed and linked.

Change-Id: Ib569a46f00866bdaa4017f803d85f625ae011316