cosmetic: clarify language in vty read error
This is very minor but it annoys every time I see it.
The text: "Error occurred during reading below line:"
is not a complete sentence. The default understanding
in english having left out the article implies
that the error occured reading below [the] specified line, not
that the error occured reading [the] specified line.

That is to say, The message implied that the printed line
was the last successfully parsed line.

Change-Id: Ib4dd135feb9609b14983db5dac321a70267d8f30